Changed Character BP to 2DChar
Added PaperZD Plugin and started work on changing the default 3D character to a 2D version.

New character is named BP_2DChar and some interactions have been corrected, but there are still some that need to be fixed.

Added a spritesheet from online as placeholder content.

Need to create state machine (locomotion) for the actor.