Fail on Light Puzzle
Light puzzle is 99% done, but for what reason, when trying to setup my door to trigger when [targetActor] is hit, but Unreal/or something else does not want my  target actor to connect to the door. I try doing a trigger box, didn't work either. At this point we either scrap the light puzzle, and use our other puzzles to fill the gap, if someone else wants to give it a shot go for it. If we still want some other kind of light puzzle, I can try looking for one, that is both simple and time friendly.